+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "colormap": "^2.3.2"
+  }
+}
